requirements.txt
fastapi~=0.95.0
SQLAlchemy~=2.0.8
config~=0.5.1
pydantic~=1.10.6
python-dotenv~=1.0.0
alembic~=1.10.2
psycopg2-binary